% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbours.R
\name{find_neighbours}
\alias{find_neighbours}
\title{Find local neighbours}
\usage{
find_neighbours(data)
}
\arguments{
\item{data}{A data frame containing the observations.}
}
\value{
A list containing the neighbour indices for each observation.
}
\description{
Finds the nearest neighbours for each observation using C++.
}
